+++ tests/integration/orders/test_orders_update_and_delete_endpoints.py
@@ -1,17 +1,17 @@
-# tests/integration/test_update_and_delete_products_endpoints.py
+# tests/integration/test_orders_update_and_delete_endpoints.py
 import pytest
 from fastapi import status
 from fastapi.testclient import TestClient
-from src.api.v1.apps.orders.models import Product
+from src.api.v1.apps.orders.models import Order
 
 # 1. app fixture in tests/conftest.py 
 # ==============================================================================
-# DATA SETUP FIXTURE (Authenticated for product creation)
+# DATA SETUP FIXTURE (Authenticated for order creation)
 # ==============================================================================
 @pytest.fixture(scope="function")
-def setup_product(client, authorized_client_cookies):
+def setup_order(client, authorized_client_cookies):
     """
-    Creates a baseline product in the catalogue to reuse across test cases.
+    Creates a baseline order in the catalogue to reuse across test cases.
     Injects authorized cookies to safely bypass authentication during creation.
     """
     
@@ -19,39 +19,39 @@ def setup_product(client, authorized_client_cookies):
     client.cookies.update(authorized_client_cookies)
     
     payload = {
-        "main_dish": "taco de carnitas",
-        "category": "alimentos",
-        "price": 28.50
+        "table_no": 1,
+        "number_of_persons": 2,
+        "items": []
     }
     
-    # Act: Create the product
-    response = client.post("/orders/products/", json=payload)
+    # Act: Create the order
+    response = client.post("/orders/create", json=payload)
     # Assert: the endpoint must response with a 201 response code
     assert response.status_code == status.HTTP_201_CREATED
-    # our product is ready to be updated/deleted
+    # our order is ready to be updated/deleted
     return response.json()
 
 
-class TestUpdateDeleteProducts:
+class TestUpdateDeleteOrders:
 
     # ==============================================================================
     # TEST 1: Unauthenticated User (No Cookie / Missing Credentials)
     # ==============================================================================
-    def test_products_update_unauthenticated_redirects(self,client, setup_product):
+    def test_orders_update_unauthenticated_redirects(self,client, setup_order):
         """
         Ensure that accessing the protected UI route without an active session
         or auth cookie results in a clean redirect or unauthorized handling.
         """
         
-        product_id = setup_product["id"]
+        order_id = setup_order["id"]
 
-        # Act: once product retreived, clear the cookie
+        # Act: once order retreived, clear the cookie
         client.cookies.clear()
         
         # Act: Execute the patch request
         response = client.patch(
-            f"/orders/products/{product_id}", 
-            json={"price": 32.00},
+            f"/orders/update/{order_id}", 
+            json={"number_of_persons": 3},
             follow_redirects=False
         )
         
@@ -60,9 +60,9 @@ class TestUpdateDeleteProducts:
         assert response.headers["location"] == "/auth/login"
     
     # ==============================================================================
-    # TEST 2: test the PATCH orders/products/{id} enpoint 
+    # TEST 2: test the PATCH orders/{id} enpoint 
     # ==============================================================================
-    def test_products_update(self, client, authorized_client_cookies,setup_product, session):
+    def test_orders_update(self, client, authorized_client_cookies,setup_order, session):
         """
         Ensure that accessing the protected UI route without an active session
         or auth cookie results in a clean redirect or unauthorized handling.
@@ -70,49 +70,49 @@ class TestUpdateDeleteProducts:
         # Act: creates an authorized user
         client.cookies.update(authorized_client_cookies)
         
-        # Act: retrieve the product id
-        product_id = setup_product["id"]
+        # Act: retrieve the order id
+        order_id = setup_order["id"]
         
         # Act: Execute the patch request
         response = client.patch(
-            f"/orders/products/{product_id}", 
-            json={"price": 32.00},
+            f"/orders/update/{order_id}", 
+            json={"number_of_persons": 3},
             follow_redirects=False
         )
         data = response.json()
         # Assert:  Verify that response code is a 200 OK
         assert response.status_code == status.HTTP_200_OK
-        # Assert: Verify that the price was successfully changed
-        assert float(data["price"]) == 32.00 # response sends strings types
+        # Assert: Verify that the number of persons was successfully changed
+        assert int(data["number_of_persons"]) == 3
         # Act: clean the session for database updated veryfication
         session.expire_all()
         session.rollback()
         # Act: creates a db object
-        db_product = session.get(Product, product_id)
+        db_order = session.get(Order, order_id)
         # Assert: query executed successfully
-        assert db_product is not None
+        assert db_order is not None
         # Assert: Verify the information in database
-        assert float(db_product.price) == 32.00
+        assert float(db_order.number_of_persons) == 3
         # Assert: Verify that others fields remains untouched
-        assert db_product.main_dish == "taco de carnitas"
+        assert db_order.table_no == 1
 
     # ==============================================================================
     # TEST 3: Unauthenticated User (No Cookie / Missing Credentials) for delete endpoint
     # ==============================================================================
-    def test_products_delete_unauthenticated_redirects(self, client, setup_product):
+    def test_orders_delete_unauthenticated_redirects(self, client, setup_order):
         """
         Ensure that accessing the protected UI route without an active session
         or auth cookie results in a clean redirect or unauthorized handling.
         """
         
-        product_id = setup_product["id"]
+        order_id = setup_order["id"]
 
-        # Act: once product retreived, clear the cookie
+        # Act: once order retreived, clear the cookie
         client.cookies.clear()
         
         # Act: Execute the patch request
         response = client.delete(
-            f"/orders/products/{product_id}", 
+            f"/orders/delete/{order_id}", 
             follow_redirects=False
         )
         
@@ -121,9 +121,9 @@ class TestUpdateDeleteProducts:
         assert response.headers["location"] == "/auth/login"
     
     # ==============================================================================
-    # TEST 4: test the DELETE orders/products/{id} enpoint 
+    # TEST 4: test the DELETE orders/{id} enpoint 
     # ==============================================================================
-    def test_products_delete(self, client, authorized_client_cookies,setup_product, session):
+    def test_orders_delete(self, client, authorized_client_cookies,setup_order, session):
         """
         Ensure that accessing the protected UI route without an active session
         or auth cookie results in a clean redirect or unauthorized handling.
@@ -131,26 +131,27 @@ class TestUpdateDeleteProducts:
         # Act: creates an authorized user
         client.cookies.update(authorized_client_cookies)
         
-        # Act: retrieve the product id
-        product_id = setup_product["id"]
+        # Act: retrieve the order id
+        order_id = setup_order["id"]
 
-        # Assert that the product exists
-        assert product_id is not None
+        # Assert that the order exists
+        assert order_id is not None
         
         # Act: Execute the delete request
         response = client.delete(
-            f"/orders/products/{product_id}", 
+            f"/orders/delete/{order_id}", 
             follow_redirects=False
         )
         
         # Assert:  Verify that response is succesfully
-        assert response.status_code == 200
+        assert response.status_code == 204
         
         # Act: clean the session for database updated veryfication
         session.expire_all()
+        session.rollback()
 
         # Act: creates a query of the deleted row
-        db_product = session.get(Product, product_id)
+        db_order = session.get(Order, order_id)
         # Assert: query executed successfully
-        assert db_product is None
+        assert db_order is None